Renamed two modules in zhmccli.
Details:
- renamed _cmd_client.py to _cmd_info.py, in order to reflect the
  command it supports ('info').
- renamed _cmd_helper.py to _helper.py, in order to reflect that
  it does not implement commands at all.
